requirements.txt
aiohttp==3.7.4.post0
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
base58==1.0.3
bleach==3.3.1
bravado==11.0.3
bravado-core==5.17.0
bson==0.5.10
cached-property==1.5.2
cattrs==1.7.1
cattrs-extras==0.1.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
cytoolz==0.11.0
dateutils==0.6.12
decorator==5.0.9
defusedxml==0.7.1
deprecation==2.1.0
discord.py==1.7.3
docker==4.4.4
entrypoints==0.3
eth-hash==0.3.1
eth-typing==2.2.2
eth-utils==1.10.0
fastecdsa==1.7.5
idna==3.2
importlib-metadata==4.6.1
ipykernel==5.5.5
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
jsonpointer==2.1
jsonref==0.2
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
loguru==0.5.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
mnemonic==0.19
monotonic==1.6
msgpack==1.0.2
multidict==5.1.0
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
netstruct==1.1.2
notebook==6.4.0
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
ply==3.11
prometheus-client==0.11.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
py-ecc==5.2.0
pyblake2==1.1.2
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
pysha3==1.0.2
pysodium==0.7.7
pytezos==3.2.5
python-dateutil==2.8.2
pytimeparse==1.1.8
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.26.0
rfc3987==1.3.8
secp256k1==0.13.2
Send2Trash==1.7.1
simplejson==3.17.3
six==1.16.0
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
tabulate==0.8.9
terminado==0.10.1
testcontainers==3.4.1
testpath==0.5.0
toolz==0.11.1
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
typing-extensions==3.10.0.0
typing-inspect==0.6.0
urllib3==1.26.6
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websocket-client==1.1.0
wrapt==1.12.1
yarl==1.6.3
zipp==3.5.0
